config: address Phase C #5 review findings (regex schema, file split, tests)
Tighten secret_name schema regex to match the metadata fetcher's
defense-in-depth check so invalid names fail at parse time rather than
crashing at boot inside the fetcher (finding 1). Split loader.test.ts
into loader.test.ts and loader-metadata.test.ts so neither file blows
the 300-line cap by much, with duplicated writeYaml/cleanup helpers and
distinct TEST_DIR paths to avoid races (finding 2). Document the
single as-unknown-as cast at the loader.ts header so future maintainers
know it is the only typed/generic boundary in the file (finding 3).
Add two error-path regression tests to metadata-fetcher.test.ts for
the 304-without-cache server-bug branch and the wrapped network-error
branch (finding 5).
